Add measure segments

+++ src/main/java/org/example/musicscorebuilder/components/music/Measure.java
@@ -2,8 +2,8 @@ package org.example.musicscorebuilder.components.music;
 
 public class Measure {
     private int number;
-    private double duration = 4;
+    private int duration = 4;
 
     public int getNumber() { return number; }
-    public double getDuration() { return duration; }
+    public int getDuration() { return duration; }
 }
